:recycle: api: Move certification course context to session management
