--- api/src/certification/course/routes.js
@@ -1,5 +1,0 @@
-import * as certificationCourse from './application/certification-course-route.js';
-
-const certificationCourseRoutes = [certificationCourse];
-
-export { certificationCourseRoutes };
