+++ api/src/certification/session-management/application/certification-course-controller.js
@@ -1,8 +1,8 @@
 import * as events from '../../../../lib/domain/events/index.js';
 import { usecases as usecasesSessionManagement } from '../../session-management/domain/usecases/index.js';
 import { usecases } from '../domain/usecases/index.js';
-import * as juryCommentSerializer from '../infrastructure/serializers/jsonapi/jury-comment-serializer.js';
-import * as v3CertificationDetailsForAdministrationSerializer from '../infrastructure/serializers/jsonapi/v3-certification-course-details-for-administration-serializer.js';
+import * as juryCommentSerializer from '../infrastructure/serializers/jury-comment-serializer.js';
+import * as v3CertificationDetailsForAdministrationSerializer from '../infrastructure/serializers/v3-certification-course-details-for-administration-serializer.js';
 
 const reject = async function (request, h, dependencies = { events }) {
   const certificationCourseId = request.params.id;
@@ -48,7 +48,7 @@ const getCertificationV3Details = async function (
   dependencies = { v3CertificationDetailsForAdministrationSerializer },
 ) {
   const { certificationCourseId } = request.params;
-  const certificationDetails = await usecases.getV3CertificationCourseDetailsForAdministration({
+  const certificationDetails = await usecasesSessionManagement.getV3CertificationCourseDetailsForAdministration({
     certificationCourseId,
   });
