+++ api/src/certification/session-management/application/certification-course-route.js
