+++ api/src/certification/session-management/domain/models/V3CertificationCourseDetailsForAdministration.js
