+++ api/src/certification/session-management/domain/usecases/end-assessment-by-supervisor.js
@@ -1,3 +1,4 @@
+// TO CHECK
 const endAssessmentBySupervisor = async function ({ certificationCandidateId, certificationAssessmentRepository }) {
   const certificationAssessment =
     await certificationAssessmentRepository.getByCertificationCandidateId(certificationCandidateId);
