+++ api/src/certification/session-management/domain/usecases/get-next-challenge-for-v2-certification.js
