+++ api/src/certification/session-management/domain/usecases/get-next-challenge-for-v3-certification.js
