+++ api/src/certification/session-management/domain/usecases/get-v3-certification-course-details-for-administration.js
