+++ api/src/certification/session-management/domain/usecases/reject-certification-course.js
