+++ api/src/certification/session-management/infrastructure/repositories/index.js
@@ -1,6 +1,9 @@
 import * as certificationIssueReportRepository from '../../../../certification/shared/infrastructure/repositories/certification-issue-report-repository.js';
 import * as issueReportCategoryRepository from '../../../../certification/shared/infrastructure/repositories/issue-report-category-repository.js';
+import * as answerRepository from '../../../../shared/infrastructure/repositories/answer-repository.js';
 import * as assessmentRepository from '../../../../shared/infrastructure/repositories/assessment-repository.js';
+import * as challengeRepository from '../../../../shared/infrastructure/repositories/challenge-repository.js';
+import * as competenceRepository from '../../../../shared/infrastructure/repositories/competence-repository.js';
 import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
 import * as certificationChallengeLiveAlertRepository from '../../../shared/infrastructure/repositories/certification-challenge-live-alert-repository.js';
 import * as certificationCourseRepository from '../../../shared/infrastructure/repositories/certification-course-repository.js';
@@ -14,15 +17,16 @@ import * as sessionForSupervisingRepository from './session-for-supervising-repo
 import * as sessionJuryCommentRepository from './session-jury-comment-repository.js';
 import * as sessionRepository from './session-repository.js';
 import * as supervisorAccessRepository from './supervisor-access-repository.js';
+import * as v3CertificationCourseDetailsForAdministrationRepository from './v3-certification-course-details-for-administration-repository.js';
 
 /**
  * Using {@link https://jsdoc.app/tags-type "Closure Compiler's syntax"} to document injected dependencies
  *
- * @typedef {assessmentRepository} AssessmentRepository
  * @typedef {assessmentResultJuryCommentRepository} AssessmentResultJuryCommentRepository
  * @typedef {certificationCourseRepository} CertificationCourseRepository
  * @typedef {certificationChallengeLiveAlertRepository} CertificationChallengeLiveAlertRepository
  * @typedef {certificationOfficerRepository} CertificationOfficerRepository
+ * @typedef {competenceRepository} CompetenceRepository
  * @typedef {finalizedSessionRepository} FinalizedSessionRepository
  * @typedef {jurySessionRepository} JurySessionRepository
  * @typedef {sessionForInvigilatorKitRepository} SessionForInvigilatorKitRepository
@@ -33,13 +37,14 @@ import * as supervisorAccessRepository from './supervisor-access-repository.js';
  * @typedef {sessionRepository} SessionRepository
  * @typedef {supervisorAccessRepository} SupervisorAccessRepository
  * @typedef {certificationReportRepository} CertificationReportRepository
+ * @typedef {v3CertificationCourseDetailsForAdministrationRepository} V3CertificationCourseDetailsForAdministrationRepository
  */
 const repositoriesWithoutInjectedDependencies = {
-  assessmentRepository,
   assessmentResultJuryCommentRepository,
   certificationCourseRepository,
   certificationChallengeLiveAlertRepository,
   certificationOfficerRepository,
+  competenceRepository,
   finalizedSessionRepository,
   jurySessionRepository,
   sessionForInvigilatorKitRepository,
@@ -50,6 +55,7 @@ const repositoriesWithoutInjectedDependencies = {
   sessionJuryCommentRepository,
   supervisorAccessRepository,
   certificationReportRepository,
+  v3CertificationCourseDetailsForAdministrationRepository,
 };
 
 /**
@@ -58,4 +64,4 @@ const repositoriesWithoutInjectedDependencies = {
  */
 const dependencies = {};
 const sessionRepositories = injectDependencies(repositoriesWithoutInjectedDependencies, dependencies);
-export { assessmentRepository, sessionRepositories };
+export { answerRepository, assessmentRepository, challengeRepository, sessionRepositories };
