+++ api/src/certification/session-management/infrastructure/repositories/v3-certification-course-details-for-administration-repository.js
