+++ api/tests/certification/session-management/unit/application/certification-course-controller_test.js
@@ -1,8 +1,7 @@
 import { CertificationCourseRejected } from '../../../../../lib/domain/events/CertificationCourseRejected.js';
 import { CertificationCourseUnrejected } from '../../../../../lib/domain/events/CertificationCourseUnrejected.js';
-import { certificationCourseController } from '../../../../../src/certification/course/application/certification-course-controller.js';
-import { usecases } from '../../../../../src/certification/course/domain/usecases/index.js';
-import { usecases as sessionManagementUsecases } from '../../../../../src/certification/session-management/domain/usecases/index.js';
+import { certificationCourseController } from '../../../../../src/certification/session-management/application/certification-course-controller.js';
+import { usecases } from '../../../../../src/certification/session-management/domain/usecases/index.js';
 import { expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Controller | Certification | session-management | certification-course-controller', function () {
@@ -65,7 +64,7 @@ describe('Unit | Controller | Certification | session-management | certification
   describe('#updateJuryComment', function () {
     it('should updateJuryComment usecase', async function () {
       // given
-      sinon.stub(sessionManagementUsecases, 'updateJuryComment');
+      sinon.stub(usecases, 'updateJuryComment');
       const request = {
         auth: {
           credentials: {
@@ -83,13 +82,13 @@ describe('Unit | Controller | Certification | session-management | certification
           },
         },
       };
-      sessionManagementUsecases.updateJuryComment.resolves();
+      usecases.updateJuryComment.resolves();
 
       // when
       await certificationCourseController.updateJuryComment(request, hFake);
-      sessionManagementUsecases;
+
       // then
-      expect(sessionManagementUsecases.updateJuryComment).to.have.been.calledWithExactly({
+      expect(usecases.updateJuryComment).to.have.been.calledWithExactly({
         certificationCourseId: 123,
         assessmentResultCommentByJury: 'Tell me why',
         juryId: request.auth.credentials.userId,
