+++ api/tests/certification/session-management/unit/domain/models/V3CertificationCourseDetailsForAdministration_test.js
