+++ api/tests/certification/session-management/unit/domain/usecases/create-certification-challenge-live-alert_test.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 
-import { createCertificationChallengeLiveAlert } from '../../../../../../src/certification/course/domain/usecases/create-certification-challenge-live-alert.js';
+import { createCertificationChallengeLiveAlert } from '../../../../../../src/certification/session-management/domain/usecases/create-certification-challenge-live-alert.js';
 import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | create-certification-challenge-live-alert', function () {
