+++ api/tests/certification/session-management/unit/domain/usecases/get-next-challenge-for-v3-certification_test.js
@@ -1,5 +1,5 @@
 import { AssessmentEndedError } from '../../../../../../lib/domain/errors.js';
-import { getNextChallengeForV3Certification } from '../../../../../../src/certification/course/domain/usecases/get-next-challenge-for-v3-certification.js';
+import { getNextChallengeForV3Certification } from '../../../../../../src/certification/session-management/domain/usecases/get-next-challenge-for-v3-certification.js';
 import { CERTIFICATION_VERSIONS } from '../../../../../../src/certification/shared/domain/models/CertificationVersion.js';
 import { config } from '../../../../../../src/shared/config.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
