+++ api/tests/certification/session-management/unit/domain/usecases/get-v3-certification-course-details-for-administration_test.js
@@ -1,4 +1,4 @@
-import { getV3CertificationCourseDetailsForAdministration } from '../../../../../../src/certification/course/domain/usecases/get-v3-certification-course-details-for-administration.js';
+import { getV3CertificationCourseDetailsForAdministration } from '../../../../../../src/certification/session-management/domain/usecases/get-v3-certification-course-details-for-administration.js';
 import { AnswerStatus } from '../../../../../../src/shared/domain/models/AnswerStatus.js';
 import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
