+++ api/tests/shared/unit/application/assessments/assessment-controller_test.js
@@ -2,7 +2,7 @@ import { AssessmentCompleted } from '../../../../../lib/domain/events/Assessment
 import * as events from '../../../../../lib/domain/events/index.js';
 import { usecases } from '../../../../../lib/domain/usecases/index.js';
 import { DomainTransaction } from '../../../../../lib/infrastructure/DomainTransaction.js';
-import { usecases as certificationUsecases } from '../../../../../src/certification/course/domain/usecases/index.js';
+import { usecases as certificationUsecases } from '../../../../../src/certification/session-management/domain/usecases/index.js';
 import { usecases as devcompUsecases } from '../../../../../src/devcomp/domain/usecases/index.js';
 import { assessmentController } from '../../../../../src/shared/application/assessments/assessment-controller.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
