+++ api/tests/tooling/domain-builder/factory/build-v3-certification-challenge-for-administration.js
@@ -1,4 +1,4 @@
-import { V3CertificationChallengeForAdministration } from '../../../../src/certification/course/domain/models/V3CertificationChallengeForAdministration.js';
+import { V3CertificationChallengeForAdministration } from '../../../../src/certification/session-management/domain/models/V3CertificationChallengeForAdministration.js';
 
 export const buildV3CertificationChallengeForAdministration = ({
   challengeId,
